requirements.txt
psycopg2==2.9.10
python-dotenv==1.1.1
python-telegram-bot==22.3
jdatetime==5.2.0